DM-54342: First draft of DP2 paper (for EPD2 release)

+++ Makefile
@@ -2,9 +2,9 @@ DOCTYPE = RTN
 DOCNUMBER = 115
 DOCNAME = $(DOCTYPE)-$(DOCNUMBER)
 FLATDIR = forAAS
+SUBDIRS = figures
 
-tex = $(filter-out $(wildcard *aglossary.tex) , $(wildcard *.tex))
-
+tex = $(filter-out $(wildcard *aglossary.tex) , $(wildcard sections/*.tex) $(wildcard *.tex) )
 
 GITVERSION := $(shell git log -1 --date=short --pretty=%h)
 GITDATE := $(shell git log -1 --date=short --pretty=%ad)
@@ -15,70 +15,90 @@ endif
 
 export TEXMFHOME ?= lsst-texmf/texmf
 
-$(DOCNAME).pdf: $(tex) local.bib authors.tex aglossary.tex
+UV_RUN = uv run
+
+$(DOCNAME).pdf: $(tex) local.bib authors.tex aglossary.tex parameters_static.tex
 	latexmk -bibtex -xelatex -f $(DOCNAME)
 	makeglossaries $(DOCNAME)
 	latexmk -bibtex -xelatex -f $(DOCNAME)
 
-authors.tex:  authors.yaml
-	python3 $(TEXMFHOME)/../bin/db2authors.py > authors.tex
+authors.tex: authors.yaml
+	$(UV_RUN) python $(TEXMFHOME)/../bin/db2authors.py > authors.tex
+
+parameters_static.tex: data/static_parameters.yaml
+	$(UV_RUN) python bin/dp2_parameters.py --static-only
 
 flat:
 	if [ ! -d $(FLATDIR) ]; then \
 		mkdir $(FLATDIR) ; \
 	fi
 	latexpand --keep-comments -o $(FLATDIR)/$(DOCNAME).tex $(DOCNAME).tex
-	if [ -d "figures" ]; then \
-		cp figures/* $(FLATDIR) ;\
-	fi
+	@for dir in $(SUBDIRS); do \
+		if [ -d "$$dir" ] && [ -n "$$(ls -A $$dir 2>/dev/null)" ]; then \
+			cp $$dir/* $(FLATDIR); \
+			echo "  ✓ Copied $$dir"; \
+		fi; \
+	done
 	cp aas*.* $(FLATDIR)
 	cp *.bib $(FLATDIR)
 	cd $(FLATDIR) &&\
+	python3 $(TEXMFHOME)/../bin/extract_citations.py $(DOCNAME).tex &&\
+	sed -i '' 's|\\bibliography{local,lsst,ivoa,lsst-dm,refs_ads,refs,books}|\\bibliography{local,allcitations}|g' $(DOCNAME).tex &&\
 	latexmk -bibtex -xelatex -f $(DOCNAME) &&\
 	makeglossaries $(DOCNAME) &&\
 	latexmk -bibtex -xelatex -f $(DOCNAME) &&\
 	latexmk -c &&\
 	rm -f *.gls *.xdv *.glg *.glo *.ist *.bib &&\
-	rm README.txt
-	echo "Flat files  in $(FLATDIR)."
-
+	if [ -f README.txt ]; then rm README.txt; fi && \
+	echo "Flat files in $(FLATDIR)."
 
 .PHONY: clean
 clean:
 	latexmk -c
 	rm -f $(DOCNAME).bbl
+	rm -f $(DOCNAME).gls
 	rm -f $(DOCNAME).pdf
 	rm -f meta.tex
 	rm -f authors.tex
 	rm -f $(FLATDIR)/*
+	rm -f parameters_static.tex
 
 .FORCE:
 
+SCRIPTS_DIR = scripts
+PYTHON_SCRIPTS = $(wildcard $(SCRIPTS_DIR)/*.py)
 
-
-SCRIPTS_DIR=scripts
-PYTHON_SCRIPTS=$(wildcard $(SCRIPTS_DIR)/*.py)
-
-authors.txt:  authors.yaml
-	python3 $(TEXMFHOME)/../bin/db2authors.py -m arxiv > authors.txt
+authors.txt: authors.yaml
+	$(UV_RUN) python $(TEXMFHOME)/../bin/db2authors.py -m arxiv > authors.txt
 
 authors.csv: authors.yaml
-	python3 $(TEXMFHOME)/../bin/db2authors.py -m aascsv > authors.csv
+	$(UV_RUN) python $(TEXMFHOME)/../bin/db2authors.py -m aascsv > authors.csv
 
-aglossary.tex :$(tex) myacronyms.txt
-	python3 $(TEXMFHOME)/../bin/generateAcronyms.py -t"Sci DM Gen" -g $(tex)
+aglossary.tex: $(tex) myglossarydefs.csv
+	$(UV_RUN) python $(TEXMFHOME)/../bin/generateAcronyms.py -n -t"Sci DM" -g $(tex)
 
+.PHONY: deps
 deps:
-	pip install -r lsst-texmf/requirements.txt 
+	pip install uv
+	uv pip install -r lsst-texmf/requirements.txt
+	uv pip install -r requirements.txt
 
+# these are called by the github action and it already installs all dependanceies
 authors.yaml:
-	python3 $(TEXMFHOME)/../bin/makeAuthorListsFromGoogle.py --builder --signup 4 -p 1CGxjpPuyNJ_gXRHTvkEF0qeI0XedQ-GQgbmyzWFLSUE "RTN-115!A2:E1000"
+	python $(TEXMFHOME)/../bin/makeAuthorListsFromGoogle.py --builder --signup 4 -p 1CGxjpPuyNJ_gXRHTvkEF0qeI0XedQ-GQgbmyzWFLSUE "RTN-115!A2:E1000"
 
 skip: .FORCE
-	python3 $(TEXMFHOME)/../bin/makeAuthorListsFromGoogle.py --skip `cat skip.count` --builder --signup 4 -p 1CGxjpPuyNJ_gXRHTvkEF0qeI0XedQ-GQgbmyzWFLSUE "RTN-115!A2:E1000"
-	
+	python $(TEXMFHOME)/../bin/makeAuthorListsFromGoogle.py --skip `cat skip.count` --builder --signup 4 -p 1CGxjpPuyNJ_gXRHTvkEF0qeI0XedQ-GQgbmyzWFLSUE "RTN-115!A2:E1000"
+
+.PHONY: scripts
 scripts:
 	@echo "Running Python scripts..."
 	@for script in $(PYTHON_SCRIPTS); do \
-		python3 $$script; \
+		$(UV_RUN) python $$script; \
 	done
+.PHONY: lander
+lander:
+	uv venv --clear lander
+	. lander/bin/activate && \
+	uv pip install --python lander/bin/python -r requirements-lander.txt && \
+	lander --upload --pdf RTN-115.pdf --ltd-product rtn-115 --title "The Vera C. Rubin Observatory Data Preview 2" --handle "RTN-115" --lsstdoc "RTN-115.tex"
